refactor(typescript): Migrate custom-generator to typescript
