+++ src/utils/custom-generator.ts
@@ -1,0 +1,101 @@
+import path from 'path';
+import Generator from 'yeoman-generator';
+import { IAnswerGroup, IQuestion } from 'yeoman-generator-types';
+
+/**
+ * Custom Generator that extends Yeoman's Generator class
+ * adding extra shortcuts.
+ *
+ * It maintains a three objects with the information collected from users:
+ * - this.options: Given by yeoman, console parameters
+ * - this.defaultValues: Initially empty, may be overriden with setValue method,
+ *   values used if there is no answer and no option has been given.
+ * - this.answers: Initially empty, it is filled with the answers given
+ *   by asking questions.
+ *
+ * @see Generator
+ */
+export default class CustomGenerator extends Generator {
+  defaultValues: IAnswerGroup = {};
+  answers: IAnswerGroup = {};
+  options: IAnswerGroup;
+
+  constructor(args: any, options: any) {
+    super(args, options);
+
+    this.defaultValues = {};
+    this.answers = {};
+
+    // @ts-ignore
+    if (!this.options) {
+      this.options = {};
+    }
+  }
+
+  async _ask(question: IQuestion | IQuestion[]): Promise<IAnswerGroup> {
+    const answers = await this.prompt(question);
+    this.answers = { ...this.answers, ...answers };
+
+    return this.answers;
+  }
+
+  _copyFile(filePath: string, destinationPath: string) {
+    this.fs.copy(
+      this.templatePath(filePath),
+      this.destinationPath(destinationPath)
+    );
+  }
+
+  _copyFiles(basePath: string, filePaths: string[]) {
+    filePaths.forEach((filePath) =>
+      this._copyFile(filePath, path.join(basePath, filePath))
+    );
+  }
+
+  _copyTemplate(templatePath: string, destinationPath: string) {
+    this.fs.copyTpl(
+      this.templatePath(templatePath),
+      this.destinationPath(destinationPath),
+      { ...this.defaultValues, ...this.options, ...this.answers }
+    );
+  }
+
+  _copyTemplates(basePath: string, templatePaths: string[]) {
+    templatePaths.forEach((templatePath) =>
+      this._copyTemplate(
+        `${templatePath}.ejs`,
+        path.join(basePath, templatePath)
+      )
+    );
+  }
+
+  _getValue(key: string): string | undefined {
+    return (
+      this.answers[key] ||
+      this.options[key] ||
+      this.config.get(key) ||
+      this.defaultValues[key]
+    );
+  }
+
+  _hasValue(key: string): boolean {
+    return (
+      key in this.answers ||
+      key in this.options ||
+      Boolean(this.config.get(key)) ||
+      key in this.defaultValues
+    );
+  }
+
+  _isRequired(variable: string) {
+    const value = this._getValue(variable) || '';
+
+    if (!value || !value.trim()) {
+      this.env.error(new Error(`${variable} is required`));
+    }
+  }
+
+  _setValue(key: string, value: string) {
+    this.defaultValues[key] = value;
+  }
+}
